Extracted WithinMatcher and exec_async macro into another file
